requirements.txt
paho-mqtt==1.6.1
mongoengine==0.27.0
Pillow==9.2.0
numpy==1.21.6
moviepy==1.0.3
fastapi==0.95.2
starlette==0.27.0
pydantic==1.10.7
typing_extensions==4.4.0
